% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coo_GOYA_21gene}
\alias{coo_GOYA_21gene}
\alias{coo_GOYA_21gene_igis3}
\alias{coo_GOYA_21gene_igis4}
\title{Primary COO classifier}
\format{
An object of class 'lm'. Coefficients can be viewed by accessing
   the '$coefficients' slot.

An object of class \code{lm} of length 7.
}
\source{
Refer to the COO_Classifier.html file in inst/reports for details
on the development of this model.
}
\usage{
coo_GOYA_21gene_igis3

coo_GOYA_21gene_igis4
}
\description{
An 'lm' object containing a 21 gene COO classifier, developed from the GOYA
dataset. Two versions are available, for data processed with either
IGIS3.0 or IGIS4.0
}
\keyword{datasets}
