% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{coo_predict}
\alias{coo_predict}
\title{COO prediction function}
\usage{
coo_predict(norm_eset, model)
}
\arguments{
\item{norm_eset}{A normalized ExpressionSet, as returned by coo_normalize}

\item{model}{The fitted COO model.}
}
\value{
a matrix with three columns: 
   * Sample names
   * LPS score prediction
   * COO class
}
\description{
Function for predicting COO using a normalized RNASeq dataset
}
\details{
This function uses the model generated on the GOYA dataset to
     predict Cell of Origin on a normalized expression dataset. No subsetting
     needs to be performed, but the expression matrix must have either refseq
     (e.g. "geneID:7900") or ENSEMBL (e.g. ENSG....) gene IDs as the rownames
}
