% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{coo_rnaseq}
\alias{coo_rnaseq}
\title{Combined COO function}
\usage{
coo_rnaseq(cds)
}
\arguments{
\item{cds}{- a DESeqDataSet object, with raw counts in the "counts" slot.}
}
\value{
a matrix of COO predictions. See coo_predict for more details
}
\description{
wrapper for running coo_normalize and coo_predict together
}
\details{
This function is a wrapper for the two core COO prediction functions.
    This function intentionally forces a standard procedure for both 
    normalization and modeling, as this should give us the best chance of 
    the model matching up with our training data.
}
