% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dzsig_GAMBL_29gene}
\alias{dzsig_GAMBL_29gene}
\alias{dzsig_GAMBL_29gene_igis3}
\alias{dzsig_GAMBL_29gene_igis4}
\title{Primary DZsig classifier}
\format{
An object of class 'lm'. Coefficients can be viewed by accessing
   the '$coefficients' slot.

An object of class \code{lm} of length 12.
}
\usage{
dzsig_GAMBL_29gene_igis3

dzsig_GAMBL_29gene_igis4
}
\description{
An 'lm' object containing a 29 gene DZsig classifier, developed from the GAMBL DLBCL
dataset. Two versions are available, for data processed with either
IGIS3.0 (RefSeq IDs) or IGIS4.0 (Ensembl IDs).
}
\keyword{datasets}
