% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dzsig_predict}
\alias{dzsig_predict}
\title{DZsig prediction function}
\usage{
dzsig_predict(norm_eset, model)
}
\arguments{
\item{norm_eset}{A normalized ExpressionSet, as returned by coo_normalize}

\item{model}{The fitted DZsig model.}
}
\value{
a data frame with three columns: 
   * Sample names
   * LPS score prediction
   * DZsig class
}
\description{
Function for predicting DZsig using a normalized RNASeq dataset
}
\details{
This function uses the model generated on the GAMBL dataset to
     predict Dark Zone Signature (DZsig) on a normalized expression dataset. No subsetting
     needs to be performed, but the expression matrix must have either refseq
     (e.g. "geneID:7900") or ENSEMBL (e.g. ENSG....) gene IDs as the rownames
}
