% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dzsig_rnaseq}
\alias{dzsig_rnaseq}
\title{Combined DZsig function}
\usage{
dzsig_rnaseq(cds)
}
\arguments{
\item{cds}{- a DESeqDataSet object, with raw counts in the "counts" slot.}
}
\value{
a data frame of DZsig predictions. See dzsig_predict for more details
}
\description{
wrapper for running coo_normalize and dzsig_predict together
}
\details{
This function is a wrapper for the core normalization and prediction functions.
}
