% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_id_type}
\alias{get_id_type}
\title{Attempt to guess the identifiers used in the model}
\usage{
get_id_type(cds)
}
\arguments{
\item{cds}{- a DESeqDataSet object, with raw counts in the "counts" slot.}
}
\value{
string of IGIS version 3.0 or 4.0
}
\description{
Function for attempting to guess identifier type
}
