% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{refLibrarySize}
\alias{refLibrarySize}
\title{refLibrarySize}
\usage{
refLibrarySize(cds2, ident)
}
\arguments{
\item{cds2}{- a DESeqDataSet object, with raw counts in the "counts" slot.}

\item{ident}{- Identifiers used in the dataset.
Should be one of c("refseq","ensembl")
If not provided, the IGIS version will be inferred from the data.}
}
\value{
vector describing the robust library size for the sample
}
\description{
Function for library ref size
}
