% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{refined_coo_rnaseq}
\alias{refined_coo_rnaseq}
\title{Refined COO prediction function}
\usage{
refined_coo_rnaseq(cds)
}
\arguments{
\item{cds}{- a DESeqDataSet object, with raw counts in the "counts" slot.}
}
\value{
a data frame with six columns: 
   * Sample names
   * COO LPS score prediction
   * COO class
	  * DZsig LPS score prediction
   * DZsig class
   * Refined COO class
}
\description{
Wrapper for running coo_normalize, coo_predict, and dzsig_predict
}
\details{
This function is a wrapper that runs the normalization and both COO
   and DZsig predictions, and then combines the results into a single data frame, 
	  including the refined COO class.
}
